requirements.txt
pyJWT
psycopg2-binary
pymongo
python-dotenv
flask-cors>=3.0.6
flask-migrate==2.1.1
flask-script==2.0.6
flask-sqlalchemy==2.3.2
flask==1.1.2
celery
redis